pyproject.toml
[project]
name = "patch-icl"
version = "0.1.0"
description = "patch-based icl on images"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "nibabel>=5.4.2",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "wandb>=0.19.0",
    "hydra-core>=1.3",
]
